Effects list: the popovers paint in the popover pass, and no write past the end
The gate's remaining failures, one at a time:

- **The menus' clicks went nowhere.** `paint_panel_menus` collected its hits
  into a local vec and appended them to `app.hit` — but it ran from the overlay
  pass, which *takes* `app.hit` first and *assigns* it back at the end, so every
  popover hit was overwritten before the frame ended. The menus now paint from
  the popover pass in `paint()` (`paint_effects_menus`), with the frame,
  line-height and text-style dropdowns, which is also what lets them hang past
  the panel's edge: that pass runs after `paint_right` has popped its clip
  layer and run its viewport hit filter.
- The section's `+` menu had the same disease from the other side. It was drawn
  inside the clipped panel, so a menu anchored to a section that sits below the
  fold lost its rows to the filter — and was cut off visually. It moves into the
  popover pass too, anchored by `effect_add_anchor`, and it no longer reflows the
  panel it floats over.
- **A write past the end of the stack pushed an undo entry.**
  `remove_effect_layer` and `move_effect_layer` were the two setters without the
  `effect_at` guard, so a click on a row that is gone (or a drop index past the
  end) made a no-op history entry. Both refuse now, which is what the test
  `an_effect_write_past_the_end_changes_nothing` was already asking for.
- The stack stores `f32`, so the panel's Density round-trip is f32-accurate:
  the test asserts `f64::from(0.6_f32)` instead of `0.6`.
- `EffectKind::Noise` is a unit variant — the clippy warning (`struct pattern is
  not needed`) goes with the fix.
- An effect's Fill fallback in the colour popover is a role (`C_TEXT`), not a
  bare `Color::BLACK`: the design-token ink ceiling for `editor_ui.rs` is 3, and
  it was at 4.

The three app tests now reach the section the way a user does — `scroll_effects_
into_view` paints once to learn where the rows landed and scrolls the panel to
them — because the section is the tail of the DESIGN column and the panel drops
hit rects that leave its viewport. And the increment gets its screenshot job
entries: `editor-effects`, `editor-effects-menu` and `editor-effects-blend`
(two effects on a layer with *Effect settings* open, the five-type `+` menu, and
a shadow's blend menu — the one that has no *Pass through*).

+++ X-Native/crates/x-editor/src/tests_mod.rs
@@ -2572,7 +2572,12 @@ mod tests {
             n.effect_layers[0].effect.color(),
             Some(Color::from_rgba8(255, 0, 0, 128))
         );
-        assert_eq!(n.effect_layers[1].effect.field(EffectField::Density), 0.6);
+        // the stack holds f32, so the round-trip is f32-accurate (0.6 → the
+        // nearest f32, which is what the panel prints back as 60%)
+        assert_eq!(
+            n.effect_layers[1].effect.field(EffectField::Density),
+            f64::from(0.6_f32)
+        );
 
         // per-effect visibility: hidden, but the settings stay
         assert!(e.set_effect_layer_visible("r1", 0, false));
